+++ MaxMind.GeoIP2/Model/Country.cs
@@ -33,12 +33,22 @@ namespace MaxMind.GeoIP2.Model
         }
 
         /// <summary>
+        ///     Constructor for binary compatibility
+        /// </summary>
+        public Country(int? confidence = null, long? geoNameId = null, string isoCode = null,
+            IDictionary<string, string> names = null, IEnumerable<string> locales = null)
+            : this(confidence, geoNameId, false, isoCode, names, locales)
+        {
+        }
+
+        /// <summary>
         ///     Constructor
         /// </summary>
         [Constructor]
         public Country(
             int? confidence = null,
             [Parameter("geoname_id")] long? geoNameId = null,
+            [Parameter("is_in_european_union")] bool isInEuropeanUnion = false,
             [Parameter("iso_code")] string isoCode = null,
             IDictionary<string, string> names = null,
             IEnumerable<string> locales = null)
@@ -46,6 +56,7 @@ namespace MaxMind.GeoIP2.Model
         {
             Confidence = confidence;
             IsoCode = isoCode;
+            IsInEuropeanUnion = isInEuropeanUnion;
         }
 
         /// <summary>
@@ -55,6 +66,14 @@ namespace MaxMind.GeoIP2.Model
         /// </summary>
         [JsonProperty("confidence")]
         public int? Confidence { get; internal set; }
+
+        /// <summary>
+        ///     This is true if the country is a member state of the
+        ///     European Union. This is available from  all location
+        ///     services and databases.
+        /// </summary>
+        [JsonProperty("is_in_european_union")]
+        public bool IsInEuropeanUnion { get; internal set; }
 
         /// <summary>
         ///     The