+++ MaxMind.GeoIP2/Model/RepresentedCountry.cs
@@ -27,11 +27,20 @@ namespace MaxMind.GeoIP2.Model
         }
 
         /// <summary>
-        ///     Constructor
+        ///     Constructor for binary compatibility
         /// </summary>
         public RepresentedCountry(string type = null, int? confidence = null, int? geoNameId = null,
             string isoCode = null, IDictionary<string, string> names = null, IEnumerable<string> locales = null)
             : this(type, confidence, (long?)geoNameId, isoCode, names, locales)
+        {
+        }
+
+        /// <summary>
+        ///     Constructor for binary compatibility
+        /// </summary>
+        public RepresentedCountry(string type = null, int? confidence = null, long? geoNameId = null,
+            string isoCode = null, IDictionary<string, string> names = null, IEnumerable<string> locales = null)
+            : this(type, confidence, geoNameId, false, isoCode, names, locales)
         {
         }
 
@@ -43,10 +52,11 @@ namespace MaxMind.GeoIP2.Model
             string type = null,
             int? confidence = null,
             [Parameter("geoname_id")] long? geoNameId = null,
+            [Parameter("is_in_european_union")] bool isInEuropeanUnion = false,
             [Parameter("iso_code")] string isoCode = null,
             IDictionary<string, string> names = null,
             IEnumerable<string> locales = null)
-            : base(confidence, geoNameId, isoCode, names, locales)
+            : base(confidence, geoNameId, isInEuropeanUnion, isoCode, names, locales)
         {
             Type = type;
         }